pyproject.toml
[project]
name = "compile-wiki"
version = "0.2.0"
description = "Compile: an LLM-maintained research wiki. Drop sources, get a living knowledge base."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "click>=8.1",
  "httpx>=0.28.1",
  "pyyaml>=6.0.3",
  "rich>=13.0",
  "watchdog>=4.0",
]

[project.scripts]
compile = "compile.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["compile"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
